Add files via upload
Design a system for the management of flights and pilots.

Airlines operate the flights. Every airline has an identity.
The airline has different types of aircraft.
Aircraft may be in operation or in a state of repair.
Each flight has a unique identity, airport to take off and land at, and departure and landing times.
Every flight has a pilot and co-pilot, and they operate the plane.
Airports have unique IDs and names.
Airlines have pilots, and each pilot has a level of experience.
An aircraft type may need a certain number of pilots.
Draw the Class diagram describing this system.